+++ src/cli/commands/connected-accounts.ts
@@ -1,0 +1,33 @@
+import executeCommand from "../execute-command"
+import { YargsWithGlobalOptions } from "../global-options"
+
+export default {
+  command: "connected-accounts",
+  aliases: ["ca", "connected-account"],
+  describe: "interact with connected accounts",
+  builder: (yargs: YargsWithGlobalOptions) => {
+    yargs
+      .demandCommand()
+      .command(
+        "list",
+        "list connected accounts",
+        () => {},
+        async (argv) => {
+          await executeCommand("connectedAccounts.list", [], argv)
+        }
+      )
+      .command(
+        "get <id>",
+        "get a connected account",
+        (yargs) => {
+          return yargs.positional("id", {
+            describe: "the connected account ID",
+            demandOption: true,
+          })
+        },
+        async (argv) => {
+          await executeCommand("connectedAccounts.get", [argv.id], argv)
+        }
+      )
+  },
+}
