+++ src/cli/index.ts
@@ -1,0 +1,14 @@
+#!/usr/bin/env node
+import commonOptions from "./global-options"
+
+// todo: bash auto-completion
+
+commonOptions
+  .commandDir("./commands", {
+    extensions: ["ts"],
+    visit: (command) => {
+      return command.default
+    },
+  })
+  .demandCommand()
+  .parse()
